requirements.txt
pandas
openpyxl
mariadb
pyyaml
python-docx
jinja2
plotly
kaleido
graphviz
networkx
matplotlib